ai_anomaly/requirements.txt

psycopg2-binary==2.9.9
scikit-learn==1.3.2
numpy==1.26.0
pandas==2.1.1
python-dotenv==1.0.1
